[blog] "Fix" the sidebar showing nanobar README.md
Jekyll loops through all pages and shows them on the sidebar. .md files
count as pages. Since I imported nanobar it shows up as a page which is
weird. I don't know where the correct place to put the asset is so
instead I am going to hotfix this by removing the README.md